Export: Refactor Export and GT components
This arranges the export and guided transfer component files
into a heirarchy more closely matching the way they are
displayed on the page, and improves separation of concerns.

Prior to this change, the export component and guided transfer
state were intermingled with a long list of props. This separates the
two and adds Redux connectors where necessary to ensure each card
only has access to the state it requires.